requirements.txt
streamlit
tensorflow
pillow
numpy
requests
pandas
scikit-learn
plotly
python-dotenv
pydub
gtts
googletrans==3.1.0a0
google-generativeai
google-genai
bcrypt
fpdf2
folium
streamlit_folium
